Remove HttpClient class
Execute the parse logic in background tasks. Update the UI in the
onPostExecute callback.

+++ WeatherApp/app/src/main/java/com/example/weatherapp/model/MainDTO.java
@@ -10,9 +10,9 @@ public class MainDTO {
 
     public float temp;
 
-    public int pressure;
+    public float pressure;
 
-    public int humidity;
+    public float humidity;
 
     @SerializedName("temp_min")
     public float tempMin;
